refactor(ui): 重构UI层级管理系统
- 移除 System.Collections.Generic 和 System 的 using 语句
- 添加 GFramework.Game.Abstractions.enums 的 using 语句
- 添加静态只读字典 LayerZOrderMap 来管理 UI 层级 Z 轴顺序
- 修改 AddUiPage 方法参数，使用 UiLayer 枚举替代 zOrder 参数
- 使用 TryAdd 方法替换 ContainsKey 判断逻辑
- 使用 Remove 方法的重载版本来简化页面移除逻辑
- 删除 RefreshLayerOrder 方法
- 添加 GetBaseZOrder 辅助方法来获取基础 Z 轴顺序
- 更新 UiLayer 枚举定义，移除显式的数值赋值
- 更新 IUiRoot 接口中的 AddUiPage 方法签名
- 删除 IUiRoot 接口中 RefreshLayerOrder 方法
- 更新 UiRouterBase 中调用 AddUiPage 方法的方式，传递 UiLayer 参数

+++ GFramework.Game/ui/UiRouterBase.cs
@@ -504,7 +504,7 @@ public abstract class UiRouterBase : AbstractSystem, IUiRouter
         layerDict[uiKey] = page;
 
         // 添加到UiRoot，传入层级Z-order
-        _uiRoot.AddUiPage(page, (int)layer);
+        _uiRoot.AddUiPage(page, layer);
 
         page.OnEnter(param);
         page.OnShow();
@@ -526,7 +526,7 @@ public abstract class UiRouterBase : AbstractSystem, IUiRouter
             _layers[layer] = new Dictionary<string, IUiPageBehavior>();
 
         _layers[layer][uiKey] = page;
-        _uiRoot.AddUiPage(page, (int)layer);
+        _uiRoot.AddUiPage(page, layer);
         page.OnShow();
 
         Log.Debug("Show existing UI instance in layer: {0}, layer={1}", uiKey, layer);
